最终修改

+++ mengxuegu-security-web/src/main/java/com/mengxuegu/web/entities/SysRole.java
@@ -12,16 +12,13 @@ import java.util.Date;
 import java.util.List;
 
 /**
- * Created by Y_Coffee on 2020/9/14
- *
- * @author CoffeeY
- * Serializable 实现序列化，为了远程数据传输，不然会报错
+ * @Auther: 梦学谷 www.mengxuegu.com
  */
 @Data
-public class SysRole implements Serializable {
+public class SysRole  implements Serializable {
 
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
     /**
      * 角色名称
      */
@@ -40,19 +37,17 @@ public class SysRole implements Serializable {
      */
     @TableField(exist = false)
     private List<SysPermission> perList = Lists.newArrayList();
-
     /**
-     * 查询当前角色所拥有的全部权限，放入当前list集合
      * 存储当前角色的权限资源ID集合
      * 修改角色时用到
      */
     @TableField(exist = false)
-    private List<Integer> perIds = Lists.newArrayList();
+    private List<Long> perIds = Lists.newArrayList();
 
-    public List<Integer> getPerIds() {
-        if (CollectionUtils.isNotEmpty(perList)) {
+    public List<Long> getPerIds() {
+        if(CollectionUtils.isNotEmpty(perList)) {
             perIds = Lists.newArrayList();
-            for (SysPermission per : perList) {
+            for(SysPermission per : perList) {
                 perIds.add(per.getId());
             }
         }
